[refactor] Made a commands and options directory
