+++ mod.ts
@@ -1,6 +1,6 @@
 import { purposes } from "./dmm.ts";
 import { colours } from "./deps.ts";
-import { helpMessage } from "./src/help.ts";
+import { helpMessage } from "./src/options/help.ts";
 
 // Gather args
 const args: string[] = Deno.args;
