+++ src/options/help.ts
